Added a menu and a game over screen
Added an initial menu which only allows the player to play the game at the moment. Menu allows navigation with the selected choice highlighted by a cursor and supports dynamic layout based on the length of the array of options within the menu

Added scrolling "Game Over"  text whenever player crashes.

Change putImageV2 to allow for horizontal and vertical orientation